requirements.txt
python-dotenv

pandas
numpy

scikit-learn
matplotlib

certifi
pymongo

pymongo[srv]

pyaml

mlflow